refactor: add DL model class (#722)
* refactor: add DL model class


* fix model url in example doc


* address comments


* fix failed ut


---------
