+++ common/src/main/java/org/opensearch/ml/common/FunctionName.java
@@ -25,4 +25,15 @@ public enum FunctionName {
             throw new IllegalArgumentException("Wrong function name");
         }
     }
+
+    /**
+     * Check if model is deep learning model.
+     * @return true for deep learning model.
+     */
+    public static boolean isDLModel(FunctionName functionName) {
+        if (functionName == TEXT_EMBEDDING) {
+            return true;
+        }
+        return false;
+    }
 }
