+++ ml-algorithms/src/main/java/org/opensearch/ml/engine/algorithms/text_embedding/TextEmbeddingModel.java
@@ -5,315 +5,111 @@
 
 package org.opensearch.ml.engine.algorithms.text_embedding;
 
-import ai.djl.Application;
-import ai.djl.Device;
-import ai.djl.engine.Engine;
 import ai.djl.inference.Predictor;
 import ai.djl.modality.Input;
 import ai.djl.modality.Output;
-import ai.djl.repository.zoo.Criteria;
-import ai.djl.repository.zoo.ZooModel;
-import ai.djl.util.ZipUtils;
+import ai.djl.translate.TranslateException;
+import ai.djl.translate.Translator;
+import ai.djl.translate.TranslatorFactory;
 import lombok.extern.log4j.Log4j2;
-import org.apache.commons.io.FileUtils;
 import org.opensearch.ml.common.FunctionName;
-import org.opensearch.ml.common.MLModel;
 import org.opensearch.ml.common.dataset.MLInputDataset;
 import org.opensearch.ml.common.dataset.TextDocsInputDataSet;
-import org.opensearch.ml.common.exception.MLException;
-import org.opensearch.ml.common.input.MLInput;
 import org.opensearch.ml.common.model.MLModelConfig;
-import org.opensearch.ml.common.model.MLModelFormat;
 import org.opensearch.ml.common.model.TextEmbeddingModelConfig;
-import org.opensearch.ml.common.output.MLOutput;
 import org.opensearch.ml.common.output.model.ModelResultFilter;
 import org.opensearch.ml.common.output.model.ModelTensorOutput;
 import org.opensearch.ml.common.output.model.ModelTensors;
-import org.opensearch.ml.engine.MLEngine;
-import org.opensearch.ml.engine.ModelHelper;
-import org.opensearch.ml.engine.Predictable;
+import org.opensearch.ml.engine.algorithms.DLModel;
 import org.opensearch.ml.engine.annotation.Function;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.nio.file.Path;
-import java.security.AccessController;
-import java.security.PrivilegedActionException;
-import java.security.PrivilegedExceptionAction;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.atomic.AtomicInteger;
 
 import static org.opensearch.ml.common.model.TextEmbeddingModelConfig.FrameworkType.SENTENCE_TRANSFORMERS;
-import static org.opensearch.ml.engine.ModelHelper.ONNX_FILE_EXTENSION;
 import static org.opensearch.ml.engine.ModelHelper.ONNX_ENGINE;
-import static org.opensearch.ml.engine.ModelHelper.PYTORCH_FILE_EXTENSION;
 import static org.opensearch.ml.engine.ModelHelper.PYTORCH_ENGINE;
-import static org.opensearch.ml.engine.utils.FileUtils.deleteFileQuietly;
 
 @Log4j2
 @Function(FunctionName.TEXT_EMBEDDING)
-public class TextEmbeddingModel implements Predictable {
+public class TextEmbeddingModel extends DLModel {
 
     public static final String SENTENCE_EMBEDDING = "sentence_embedding";
-    public static final String MODEL_ZIP_FILE = "model_zip_file";
-    public static final String MODEL_HELPER = "model_helper";
-    public static final String ML_ENGINE = "ml_engine";
-
-    private ModelHelper modelHelper;
-    private MLEngine mlEngine;
-    private String modelId;
-
-    private Predictor<Input, Output>[] predictors;
-    private ZooModel[] models;
-    private Device[] devices;
-    private AtomicInteger nextDevice = new AtomicInteger(0);
 
     @Override
-    public MLOutput predict(MLInput mlInput, MLModel model) {
-        throw new MLException("model not loaded");
+    public ModelTensorOutput predict(String modelId, MLInputDataset inputDataSet) throws TranslateException {
+        List<ModelTensors> tensorOutputs = new ArrayList<>();
+        Output output;
+        TextDocsInputDataSet textDocsInput = (TextDocsInputDataSet) inputDataSet;
+        ModelResultFilter resultFilter = textDocsInput.getResultFilter();
+        for (String doc : textDocsInput.getDocs()) {
+            Input input = new Input();
+            input.add(doc);
+            output = getPredictor().predict(input);
+            tensorOutputs.add(parseModelTensorOutput(output, resultFilter));
+        }
+        return new ModelTensorOutput(tensorOutputs);
     }
 
     @Override
-    public MLOutput predict(MLInput mlInput) {
-        if (modelHelper == null || modelId == null) {
-            throw new MLException("model not loaded");
+    public Translator<Input, Output> getTranslator(String engine, MLModelConfig modelConfig) {
+        TextEmbeddingModelConfig textEmbeddingModelConfig = (TextEmbeddingModelConfig) modelConfig;
+        TextEmbeddingModelConfig.FrameworkType transformersType = textEmbeddingModelConfig.getFrameworkType();
+        String modelType = textEmbeddingModelConfig.getModelType();
+        TextEmbeddingModelConfig.PoolingMode poolingMode = textEmbeddingModelConfig.getPoolingMode();
+        boolean normalizeResult = textEmbeddingModelConfig.isNormalizeResult();
+
+        if (ONNX_ENGINE.equals(engine)) { //ONNX
+            return new ONNXSentenceTransformerTextEmbeddingTranslator(poolingMode, normalizeResult, modelType);
+        } else if (transformersType == SENTENCE_TRANSFORMERS) {// pytorch sentence_transformer
+            return new SentenceTransformerTextEmbeddingTranslator();
         }
-        return predictTextEmbedding(modelId, mlInput.getInputDataset());
+        return null;
     }
 
     @Override
-    public void initModel(MLModel model, Map<String, Object> params) {
-        String engine = model.getModelFormat() == MLModelFormat.TORCH_SCRIPT ? PYTORCH_ENGINE : ONNX_ENGINE;
-        File modelZipFile = (File)params.get(MODEL_ZIP_FILE);
-        modelHelper = (ModelHelper)params.get(MODEL_HELPER);
-        mlEngine = (MLEngine)params.get(ML_ENGINE);
-        if (modelZipFile == null) {
-            throw new IllegalArgumentException("model file is null");
+    public TranslatorFactory getTranslatorFactory(String engine, MLModelConfig modelConfig) {
+        TextEmbeddingModelConfig textEmbeddingModelConfig = (TextEmbeddingModelConfig) modelConfig;
+        TextEmbeddingModelConfig.FrameworkType transformersType = textEmbeddingModelConfig.getFrameworkType();
+        String modelType = textEmbeddingModelConfig.getModelType();
+        TextEmbeddingModelConfig.PoolingMode poolingMode = textEmbeddingModelConfig.getPoolingMode();
+        boolean normalizeResult = textEmbeddingModelConfig.isNormalizeResult();
+
+        if (PYTORCH_ENGINE.equals(engine) && transformersType != SENTENCE_TRANSFORMERS) { //pytorch
+            boolean neuron = false;
+            if (transformersType.name().endsWith("_NEURON")) {
+                neuron = true;
+            }
+            return new HuggingfaceTextEmbeddingTranslatorFactory(poolingMode, normalizeResult, modelType, neuron);
         }
-        if (modelHelper == null) {
-            throw new IllegalArgumentException("model helper is null");
-        }
-        if (mlEngine == null) {
-            throw new IllegalArgumentException("ML engine is null");
-        }
-        modelId = model.getModelId();
-        if (modelId == null) {
-            throw new IllegalArgumentException("model id is null");
-        }
-        loadTextEmbeddingModel(
-                modelZipFile,
-                modelId,
-                model.getName(),
-                model.getAlgorithm(),
-                model.getVersion(),
-                model.getModelConfig(),
-                engine
-        );
+        return null;
     }
 
     @Override
-    public void close() {
-        if (modelHelper != null && modelId != null) {
-            modelHelper.deleteFileCache(modelId);
-            if (predictors != null) {
-                closePredictors(predictors);
-                predictors = null;
-            }
-            if (models != null) {
-                closeModels(models);
-                models = null;
-            }
+    public Map<String, Object> getArguments(MLModelConfig modelConfig) {
+        TextEmbeddingModelConfig textEmbeddingModelConfig = (TextEmbeddingModelConfig) modelConfig;
+        Integer modelMaxLength = textEmbeddingModelConfig.getModelMaxLength();
+        Map<String, Object> arguments = new HashMap<>();
+        if (modelMaxLength != null) {
+            arguments.put("modelMaxLength", modelMaxLength);
         }
+        return arguments;
     }
 
-    protected void loadTextEmbeddingModel(File modelZipFile, String modelId, String modelName, FunctionName functionName, String version,
-                                       MLModelConfig modelConfig,
-                                       String engine) {
-
-        try {
-            if (FunctionName.TEXT_EMBEDDING != functionName) {
-                throw new IllegalArgumentException("wrong function name");
-            }
-            if (!PYTORCH_ENGINE.equals(engine) && !ONNX_ENGINE.equals(engine)) {
-                throw new IllegalArgumentException("unsupported engine");
-            }
-            List<Predictor<Input, Output>> predictorList = new ArrayList<>();
-            List<ZooModel<Input, Output>> modelList = new ArrayList<>();
-            AccessController.doPrivileged((PrivilegedExceptionAction<Void>) () -> {
-                ClassLoader contextClassLoader = Thread.currentThread().getContextClassLoader();
-                try {
-                    System.setProperty("PYTORCH_PRECXX11", "true");
-                    System.setProperty("DJL_CACHE_DIR", mlEngine.getDjlCachePath().toAbsolutePath().toString());
-                    // DJL will read "/usr/java/packages/lib" if don't set "java.library.path". That will throw
-                    // access denied exception
-                    System.setProperty("java.library.path", mlEngine.getDjlCachePath().toAbsolutePath().toString());
-                    System.setProperty("ai.djl.pytorch.num_interop_threads", "1");
-                    System.setProperty("ai.djl.pytorch.num_threads", "1");
-                    Thread.currentThread().setContextClassLoader(ai.djl.Model.class.getClassLoader());
-                    Path modelPath = mlEngine.getModelCachePath(modelId, modelName, version);
-                    File pathFile = new File(modelPath.toUri());
-                    if (pathFile.exists()) {
-                        FileUtils.deleteDirectory(pathFile);
-                    }
-                    try (FileInputStream fileInputStream = new FileInputStream(modelZipFile)) {
-                        ZipUtils.unzip(fileInputStream, modelPath);
-                    }
-                    boolean findModelFile = false;
-                    for (File file : pathFile.listFiles()) {
-                        String name = file.getName();
-                        if (name.endsWith(PYTORCH_FILE_EXTENSION) || name.endsWith(ONNX_FILE_EXTENSION)) {
-                            if (findModelFile) {
-                                throw new IllegalArgumentException("found multiple models");
-                            }
-                            findModelFile = true;
-                            int dotIndex = name.lastIndexOf(".");
-                            String suffix = name.substring(dotIndex);
-                            String targetModelFileName = modelPath.getFileName().toString();
-                            if (!targetModelFileName.equals(name.substring(0, dotIndex))) {
-                                file.renameTo(new File(modelPath.resolve(targetModelFileName + suffix).toUri()));
-                            }
-                        }
-                    }
-                    devices = Engine.getEngine(engine).getDevices();
-                    for (int i = 0; i < devices.length; i++) {
-                        log.debug("load model {} on device {}: {}", modelId, i, devices[i]);
-                        Criteria.Builder<Input, Output> criteriaBuilder = Criteria.builder()
-                                .setTypes(Input.class, Output.class)
-                                .optApplication(Application.UNDEFINED)
-                                .optEngine(engine)
-                                .optDevice(devices[i])
-                                .optModelPath(modelPath);
-                        TextEmbeddingModelConfig textEmbeddingModelConfig = (TextEmbeddingModelConfig) modelConfig;
-                        TextEmbeddingModelConfig.FrameworkType transformersType = textEmbeddingModelConfig.getFrameworkType();
-                        String modelType = textEmbeddingModelConfig.getModelType();
-                        TextEmbeddingModelConfig.PoolingMode poolingMode = textEmbeddingModelConfig.getPoolingMode();
-                        boolean normalizeResult = textEmbeddingModelConfig.isNormalizeResult();
-                        Integer modelMaxLength = textEmbeddingModelConfig.getModelMaxLength();
-                        if (modelMaxLength != null) {
-                            criteriaBuilder.optArgument("modelMaxLength", modelMaxLength);
-                        }
-                        //TODO: refactor this when we support more engine type
-                        if (ONNX_ENGINE.equals(engine)) { //ONNX
-                            criteriaBuilder.optTranslator(new ONNXSentenceTransformerTextEmbeddingTranslator(poolingMode, normalizeResult, modelType));
-                        } else { // pytorch
-                            if (transformersType == SENTENCE_TRANSFORMERS) {
-                                criteriaBuilder.optTranslator(new SentenceTransformerTextEmbeddingTranslator());
-                            } else {
-                                boolean neuron = false;
-                                if (transformersType.name().endsWith("_NEURON")) {
-                                    neuron = true;
-                                }
-                                criteriaBuilder.optTranslatorFactory(new HuggingfaceTextEmbeddingTranslatorFactory(poolingMode, normalizeResult, modelType, neuron));
-                            }
-                        }
-                        Criteria<Input, Output> criteria = criteriaBuilder.build();
-                        ZooModel<Input, Output> model = criteria.loadModel();
-                        Predictor<Input, Output> predictor = model.newPredictor();
-                        predictorList.add(predictor);
-                        modelList.add(model);
-
-                        Input input = new Input();
-                        if (modelMaxLength != null) {
-                            input.add("sentence ".repeat(modelMaxLength));
-                        } else {
-                            input.add("warm up sentence");
-                        }
-                        // First request takes longer time. Predict once to warm up model.
-                        predictor.predict(input);
-                    }
-                    if (predictorList.size() > 0) {
-                        this.predictors = predictorList.toArray(new Predictor[0]);
-                        predictorList.clear();
-                    }
-                    if (modelList.size() > 0) {
-                        this.models = modelList.toArray(new ZooModel[0]);
-                        modelList.clear();
-                    }
-                    log.info("Load model {} successfully on {} devices", modelId, devices.length);
-                    return null;
-                } catch (Throwable e) {
-                    String errorMessage = "Failed to load model " + modelId;
-                    log.error(errorMessage, e);
-                    close();
-                    if (predictorList.size() > 0) {
-                        closePredictors(predictorList.toArray(new Predictor[0]));
-                        predictorList.clear();
-                    }
-                    if (modelList.size() > 0) {
-                        closeModels(modelList.toArray(new ZooModel[0]));
-                        modelList.clear();
-                    }
-                    throw new MLException(errorMessage, e);
-                } finally {
-                    deleteFileQuietly(mlEngine.getLoadModelPath(modelId));
-                    Thread.currentThread().setContextClassLoader(contextClassLoader);
-                }
-            });
-        } catch (PrivilegedActionException e) {
-            String errorMsg = "Failed to load model";
-            log.error(errorMsg, e);
-            throw new MLException(errorMsg, e);
+    @Override
+    public void warmUp(Predictor predictor, String modelId, MLModelConfig modelConfig) throws TranslateException {
+        TextEmbeddingModelConfig textEmbeddingModelConfig = (TextEmbeddingModelConfig) modelConfig;
+        Integer modelMaxLength = textEmbeddingModelConfig.getModelMaxLength();
+        String warmUpSentence = "warm up sentence";
+        if (modelMaxLength != null) {
+            warmUpSentence = "sentence ".repeat(modelMaxLength);
         }
-    }
-
-    private void closePredictors(Predictor[] predictors) {
-        log.debug("will close {} predictor for model {}", predictors.length, modelId);
-        for (Predictor<Input, Output> predictor : predictors) {
-            predictor.close();
-        }
-    }
-
-    private void closeModels(ZooModel[] models) {
-        log.debug("will close {} zoo model for model {}", models.length, modelId);
-        for (ZooModel model : models) {
-            model.close();
-        }
-    }
-
-    protected ModelTensorOutput predictTextEmbedding(String modelId, MLInputDataset inputDataSet) {
-        try {
-            return AccessController.doPrivileged((PrivilegedExceptionAction<ModelTensorOutput>) () -> {
-                Thread.currentThread().setContextClassLoader(getClass().getClassLoader());
-                int currentDevice = nextDevice.getAndIncrement();
-                if (currentDevice > devices.length - 1) {
-                    currentDevice = currentDevice % devices.length;
-                    nextDevice.set(currentDevice + 1);
-                }
-                if (predictors == null) {
-                    throw new MLException("model not loaded.");
-                }
-                List<ModelTensors> tensorOutputs = new ArrayList<>();
-                Output output;
-                TextDocsInputDataSet textDocsInput = (TextDocsInputDataSet) inputDataSet;
-                ModelResultFilter resultFilter = textDocsInput.getResultFilter();
-                for (String doc : textDocsInput.getDocs()) {
-                    Input input = new Input();
-                    input.add(doc);
-                    log.debug("run text embedding predict for model {} on device {}", modelId, devices[currentDevice]);
-                    output = predictors[currentDevice].predict(input);
-                    tensorOutputs.add(parseModelTensorOutput(output, resultFilter));
-                }
-                return new ModelTensorOutput(tensorOutputs);
-            });
-        } catch (Throwable e) {
-            String errorMsg = "Failed to inference text embedding";
-            log.error(errorMsg, e);
-            throw new MLException(errorMsg, e);
-        }
-    }
-
-    protected ModelTensors parseModelTensorOutput(Output output, ModelResultFilter resultFilter) {
-        if (output == null) {
-            throw new MLException("No output generated");
-        }
-        byte[] bytes = output.getData().getAsBytes();
-        ModelTensors tensorOutput = ModelTensors.fromBytes(bytes);
-        if (resultFilter != null) {
-            tensorOutput.filter(resultFilter);
-        }
-        return tensorOutput;
+        // First request takes longer time. Predict once to warm up model.
+        Input input = new Input();
+        input.add(warmUpSentence);
+        predictor.predict(input);
     }
 
 }
